Tidy nested namespaces (#396)
