+++ include/gz/fuel_tools/Helpers.hh
@@ -35,9 +35,7 @@
   #define gz_strdup strdup
 #endif
 
-namespace gz
-{
-namespace fuel_tools
+namespace gz::fuel_tools
 {
 /// \brief Convert the authority portion of a URI to a string
 /// suitable to be used as a path on disk for all platforms.
@@ -60,8 +58,6 @@ std::string sanitizeAuthority(const std::string &_uriAuthority);
 /// \return String suitable to use in file paths
 GZ_FUEL_TOOLS_VISIBLE
 std::string uriToPath(const gz::common::URI &_uri);
-}
-}
+}  // namespace gz::fuel_tools
 
-// GZ_FUEL_TOOLS_HELPERS_HH_
-#endif
+#endif  // GZ_FUEL_TOOLS_HELPERS_HH_
