+++ include/gz/fuel_tools/Model.hh
@@ -32,89 +32,86 @@
 #pragma warning(disable: 4251)
 #endif
 
-namespace gz
+namespace gz::fuel_tools
 {
-  namespace fuel_tools
+  /// \brief Forward Declaration
+  class ModelPrivate;
+
+  /// \brief Forward declaration
+  class ModelIter;
+
+  /// \brief Forward declaration
+  class ModelIterPrivate;
+
+  /// \brief forward declaration
+  class LocalCache;
+
+  /// \brief forward declaration
+  class LocalCachePrivate;
+
+  /// \brief Forward declaration
+  class IterIds;
+
+  /// \brief Forward declaration
+  class IterRESTIds;
+
+  /// \brief Forward declaration
+  class IterRestIds;
+
+  /// \brief Forward declaration
+  class ModelIterTest;
+
+  /// \brief Defines how to identify a model.
+  class GZ_FUEL_TOOLS_VISIBLE Model
   {
-    /// \brief Forward Declaration
-    class ModelPrivate;
+    friend IterIds;
+    friend IterRESTIds;
+    friend IterRestIds;
+    friend ModelIter;
+    friend ModelIterPrivate;
+    friend ModelIterTest;
+    friend LocalCache;
+    friend LocalCachePrivate;
 
-    /// \brief Forward declaration
-    class ModelIter;
+    /// \brief Default constructor
+    public: Model();
 
-    /// \brief Forward declaration
-    class ModelIterPrivate;
+    /// \brief Copy constructor
+    /// \param[in] _orig Model to copy.
+    public: Model(const Model &_orig) = default;
 
-    /// \brief forward declaration
-    class LocalCache;
+    /// \brief Protected constructor
+    /// \param[in] _dPtr Model private data to copy.
+    protected: explicit Model(std::shared_ptr<ModelPrivate> _dptr);
 
-    /// \brief forward declaration
-    class LocalCachePrivate;
+    /// \brief Returns false if model was constructed via Model()
+    public: operator bool() const;
 
-    /// \brief Forward declaration
-    class IterIds;
+    /// \brief Returns false if model was constructed via Model().
+    /// \return False if model was constructed via Model().
+    public: operator bool();
 
-    /// \brief Forward declaration
-    class IterRESTIds;
+    /// \brief Returns information identifying the model.
+    /// \return Information that can identify the model.
+    public: ModelIdentifier Identification() const;
 
-    /// \brief Forward declaration
-    class IterRestIds;
+    /// \brief Make sure this model is in the local cache
+    /// \remarks this downloads the model and saves it locally if necessary
+    /// \returns The result of fetching
+    public: Result Fetch() const;
 
-    /// \brief Forward declaration
-    class ModelIterTest;
+    /// \brief Returns a path to the model on disk if it is already
+    /// cached.
+    /// \returns path, or empty string if the model is not cached.
+    public: std::string PathToModel() const;
 
-    /// \brief Defines how to identify a model.
-    class GZ_FUEL_TOOLS_VISIBLE Model
-    {
-      friend IterIds;
-      friend IterRESTIds;
-      friend IterRestIds;
-      friend ModelIter;
-      friend ModelIterPrivate;
-      friend ModelIterTest;
-      friend LocalCache;
-      friend LocalCachePrivate;
-
-      /// \brief Default constructor
-      public: Model();
-
-      /// \brief Copy constructor
-      /// \param[in] _orig Model to copy.
-      public: Model(const Model &_orig) = default;
-
-      /// \brief Protected constructor
-      /// \param[in] _dPtr Model private data to copy.
-      protected: explicit Model(std::shared_ptr<ModelPrivate> _dptr);
-
-      /// \brief Returns false if model was constructed via Model()
-      public: operator bool() const;
-
-      /// \brief Returns false if model was constructed via Model().
-      /// \return False if model was constructed via Model().
-      public: operator bool();
-
-      /// \brief Returns information identifying the model.
-      /// \return Information that can identify the model.
-      public: ModelIdentifier Identification() const;
-
-      /// \brief Make sure this model is in the local cache
-      /// \remarks this downloads the model and saves it locally if necessary
-      /// \returns The result of fetching
-      public: Result Fetch() const;
-
-      /// \brief Returns a path to the model on disk if it is already
-      /// cached.
-      /// \returns path, or empty string if the model is not cached.
-      public: std::string PathToModel() const;
-
-      /// \brief PIMPL
-      private: std::shared_ptr<ModelPrivate> dataPtr;
-    };
-  }
-}
+    /// \brief PIMPL
+    private: std::shared_ptr<ModelPrivate> dataPtr;
+  };
+}  // namespace gz::fuel_tools
 
 #ifdef _MSC_VER
 #pragma warning(pop)
 #endif
 
-#endif
+#endif  // GZ_FUEL_TOOLS_MODEL_HH_
