+++ include/gz/fuel_tools/Result.hh
@@ -22,120 +22,70 @@
 #include <string>
 
 #include "gz/fuel_tools/Export.hh"
+#include "gz/fuel_tools/ResultType.hh"
 
 #ifdef _WIN32
 // Disable warning C4251 which is triggered by
 // std::unique_ptr
 #pragma warning(push)
 #pragma warning(disable: 4251)
-// TODO(jrivero): rename the DELETE method which is a reserved word in Windows
-#undef DELETE
 #endif
 
-namespace gz
+namespace gz::fuel_tools
 {
-  namespace fuel_tools
+  // forward declaration
+  class ResultPrivate;
+  class Model;
+
+  /// \brief Class describing a result of an operation.
+  class GZ_FUEL_TOOLS_VISIBLE Result
   {
-    // forward declaration
-    class ResultPrivate;
-    class Model;
+    /// \brief Default constructor
+    public: Result();
 
-    /// Result type.
-    enum class ResultType
-    {
-      /// \brief Uninitialized type.
-      UNKNOWN = 0,
+    /// \brief Destructor.
+    public: virtual ~Result();
 
-      /// \brief Delete successful.
-      DELETE,
+    /// \brief Constructor
+    /// \param[in] _type Result type
+    public: explicit Result(const ResultType _type);
 
-      /// \brief Model not found.
-      DELETE_NOT_FOUND,
+    /// \brief Copy constructor.
+    /// \param[in] _result Result to copy.
+    public: Result(const Result &_result);
 
-      /// \brief Delete failed. Other errors.
-      /// \sa ReadableResult
-      DELETE_ERROR,
+    /// \brief Move constructor
+    /// \param[in] _result Result to move.
+    public: Result(Result &&_result) noexcept;  // NOLINT
 
-      /// \brief Fetch successful.
-      FETCH,
+    /// \brief Copy assignment operator.
+    /// \param[in] _result Result to copy.
+    public: Result &operator=(const Result &_result);
 
-      /// \brief Model already exists.
-      FETCH_ALREADY_EXISTS,
+    /// \brief Move assignment operator.
+    /// \param[in] _result Result component to move.
+    /// \return Reference to this.
+    public: Result &operator=(Result &&_result) noexcept;  // NOLINT
 
-      /// \brief Model not found.
-      FETCH_NOT_FOUND,
+    /// \brief Get the type of result
+    /// \return The type of result.
+    public: ResultType Type() const;
 
-      /// \brief Fetch failed. Other errors.
-      /// \sa ReadableResult
-      FETCH_ERROR,
+    /// \brief Operator bool returns true if operation was successful.
+    /// \return True if the operation was successful.
+    public: virtual operator bool() const;
 
-      /// \brief Upload successful.
-      UPLOAD,
+    /// \brief Get human readable result string.
+    /// \return The result string.
+    public: virtual std::string ReadableResult() const;
 
-      /// \brief Model already exists.
-      UPLOAD_ALREADY_EXISTS,
-
-      /// \brief Upload failed. Other errors.
-      /// \sa ReadableResult
-      UPLOAD_ERROR,
-
-      /// \brief Patch failed.
-      PATCH_ERROR,
-
-      /// \brief Patch successful.
-      PATCH,
-    };
-
-    /// \brief Class describing a result of an operation.
-    class GZ_FUEL_TOOLS_VISIBLE Result
-    {
-      /// \brief Default constructor
-      public: Result();
-
-      /// \brief Destructor.
-      public: virtual ~Result();
-
-      /// \brief Constructor
-      /// \param[in] _type Result type
-      public: explicit Result(const ResultType _type);
-
-      /// \brief Copy constructor.
-      /// \param[in] _result Result to copy.
-      public: Result(const Result &_result);
-
-      /// \brief Move constructor
-      /// \param[in] _result Result to move.
-      public: Result(Result &&_result) noexcept;  // NOLINT
-
-      /// \brief Copy assignment operator.
-      /// \param[in] _result Result to copy.
-      public: Result &operator=(const Result &_result);
-
-      /// \brief Move assignment operator.
-      /// \param[in] _result Result component to move.
-      /// \return Reference to this.
-      public: Result &operator=(Result &&_result) noexcept;  // NOLINT
-
-      /// \brief Get the type of result
-      /// \return The type of result.
-      public: ResultType Type() const;
-
-      /// \brief Operator bool returns true if operation was successful.
-      /// \return True if the operation was successful.
-      public: virtual operator bool() const;
-
-      /// \brief Get human readable result string.
-      /// \return The result string.
-      public: virtual std::string ReadableResult() const;
-
-      /// brief Pointer to private data
-      private: std::unique_ptr<ResultPrivate> dataPtr;
-    };
-  }
-}
+    /// brief Pointer to private data
+    private: std::unique_ptr<ResultPrivate> dataPtr;
+  };
+}  // namespace gz::fuel_tools
 
 #ifdef _MSC_VER
 #pragma warning(pop)
 #endif
 
-#endif
+#endif  // GZ_FUEL_TOOLS_RESULT_HH_
