+++ include/gz/fuel_tools/WorldIter.hh
@@ -29,59 +29,56 @@
 #pragma warning(disable: 4251)
 #endif
 
-namespace gz
+namespace gz::fuel_tools
 {
-  namespace fuel_tools
+  /// \brief Forward declarations
+  class WorldIdentifier;
+  class WorldIterPrivate;
+  class WorldIterFactory;
+
+  /// \brief class for iterating through worlds
+  class GZ_FUEL_TOOLS_VISIBLE WorldIter
   {
-    /// \brief Forward declarations
-    class WorldIdentifier;
-    class WorldIterPrivate;
-    class WorldIterFactory;
+    friend WorldIterFactory;
 
-    /// \brief class for iterating through worlds
-    class GZ_FUEL_TOOLS_VISIBLE WorldIter
-    {
-      friend WorldIterFactory;
+    /// \brief Construct an iterator with the data it needs to function
+    /// \param[in] _dptr Pointer to private data to copy
+    protected: explicit WorldIter(std::unique_ptr<WorldIterPrivate> _dptr);
 
-      /// \brief Construct an iterator with the data it needs to function
-      /// \param[in] _dptr Pointer to private data to copy
-      protected: explicit WorldIter(std::unique_ptr<WorldIterPrivate> _dptr);
+    /// \brief Move constructor
+    /// \param[in] _old Iter to move
+    public: WorldIter(WorldIter && _old);
 
-      /// \brief Move constructor
-      /// \param[in] _old Iter to move
-      public: WorldIter(WorldIter && _old);
+    /// \brief Default destructor.
+    public: ~WorldIter();
 
-      /// \brief Default destructor.
-      public: ~WorldIter();
+    /// \brief Conversion operator
+    /// \return False once the iterator is one past the end of the worlds
+    public: operator bool() const;
 
-      /// \brief Conversion operator
-      /// \return False once the iterator is one past the end of the worlds
-      public: operator bool() const;
+    /// \brief Conversion operator
+    /// \return Internal world identifier.
+    public: operator WorldIdentifier() const;
 
-      /// \brief Conversion operator
-      /// \return Internal world identifier.
-      public: operator WorldIdentifier() const;
+    /// \brief Prefix increment
+    /// \return Next iteration
+    public: WorldIter &operator++();
 
-      /// \brief Prefix increment
-      /// \return Next iteration
-      public: WorldIter &operator++();
+    /// \brief -> operator
+    /// \return Internal world identifier
+    public: WorldIdentifier *operator->();
 
-      /// \brief -> operator
-      /// \return Internal world identifier
-      public: WorldIdentifier *operator->();
+    /// \brief Derefence operator
+    /// \return Internal world identifier
+    public: WorldIdentifier &operator*();
 
-      /// \brief Derefence operator
-      /// \return Internal world identifier
-      public: WorldIdentifier &operator*();
-
-      /// \brief Private data pointer.
-      private: std::unique_ptr<WorldIterPrivate> dataPtr;
-    };
-  }
-}
+    /// \brief Private data pointer.
+    private: std::unique_ptr<WorldIterPrivate> dataPtr;
+  };
+}  // namespace gz::fuel_tools
 
 #ifdef _MSC_VER
 #pragma warning(pop)
 #endif
 
-#endif
+#endif  // GZ_FUEL_TOOLS_WORLDITER_HH_
