+++ src/LocalCache.hh
@@ -33,85 +33,82 @@
 #pragma warning(disable: 4251)
 #endif
 
-namespace gz
+namespace gz::fuel_tools
 {
-  namespace fuel_tools
+  /// \brief Forward declaration
+  class ClientConfig;
+  class LocalCachePrivate;
+  class ModelIdentifier;
+
+  /// \brief Class for managing stuff in the local cache
+  class GZ_FUEL_TOOLS_VISIBLE LocalCache
   {
-    /// \brief Forward declaration
-    class ClientConfig;
-    class LocalCachePrivate;
-    class ModelIdentifier;
+    /// \brief Constructor
+    /// \param[in] _config The configuration for the client
+    public: explicit LocalCache(const ClientConfig *_config);
 
-    /// \brief Class for managing stuff in the local cache
-    class GZ_FUEL_TOOLS_VISIBLE LocalCache
-    {
-      /// \brief Constructor
-      /// \param[in] _config The configuration for the client
-      public: explicit LocalCache(const ClientConfig *_config);
+    /// \brief destructor
+    public: virtual ~LocalCache();
 
-      /// \brief destructor
-      public: virtual ~LocalCache();
+    /// \brief Get all models in offline cache
+    /// \return Model iterator
+    public: virtual ModelIter AllModels();
 
-      /// \brief Get all models in offline cache
-      /// \return Model iterator
-      public: virtual ModelIter AllModels();
+    /// \brief Get all worlds in offline cache
+    /// \return World iterator
+    public: virtual WorldIter AllWorlds() const;
 
-      /// \brief Get all worlds in offline cache
-      /// \return World iterator
-      public: virtual WorldIter AllWorlds() const;
+    /// \brief Get the first model matching all fields on an id.
+    /// \param[in] _id An id with ServerUrl, Owner, and Name all set
+    /// \return A model which matches all of _id's parameters.
+    public: virtual Model MatchingModel(const ModelIdentifier &_id);
 
-      /// \brief Get the first model matching all fields on an id.
-      /// \param[in] _id An id with ServerUrl, Owner, and Name all set
-      /// \return A model which matches all of _id's parameters.
-      public: virtual Model MatchingModel(const ModelIdentifier &_id);
+    /// \brief Get the first world matching all fields on an id.
+    /// \param[in] _id An id with ServerUrl, Owner, and Name all set
+    /// \return A world which matches all of _id's parameters.
+    public: virtual bool MatchingWorld(WorldIdentifier &_id) const;
 
-      /// \brief Get the first world matching all fields on an id.
-      /// \param[in] _id An id with ServerUrl, Owner, and Name all set
-      /// \return A world which matches all of _id's parameters.
-      public: virtual bool MatchingWorld(WorldIdentifier &_id) const;
+    /// \brief Get all models partially matching an ID
+    /// \param[in] _id An id with at least one of ServerURL, Owner, and Name
+    /// \return An iterator with all models that match all fields that are
+    /// set on _id.
+    public: virtual ModelIter MatchingModels(const ModelIdentifier &_id);
 
-      /// \brief Get all models partially matching an ID
-      /// \param[in] _id An id with at least one of ServerURL, Owner, and Name
-      /// \return An iterator with all models that match all fields that are
-      /// set on _id.
-      public: virtual ModelIter MatchingModels(const ModelIdentifier &_id);
+    /// \brief Get all worlds partially matching an ID
+    /// \param[in] _id An id with at least one of ServerUrl, Owner, and Name
+    /// \return An iterator with all worlds that match all fields that are
+    /// set on _id.
+    public: virtual WorldIter MatchingWorlds(
+        const WorldIdentifier &_id) const;
 
-      /// \brief Get all worlds partially matching an ID
-      /// \param[in] _id An id with at least one of ServerUrl, Owner, and Name
-      /// \return An iterator with all worlds that match all fields that are
-      /// set on _id.
-      public: virtual WorldIter MatchingWorlds(
-          const WorldIdentifier &_id) const;
+    /// \brief Add a model from packed data to the local cache
+    /// \param[in] _id A completely populated ID
+    /// \param[in] _data Compressed content of the model
+    /// \param[in] _overwrite Overwrite model if already exists.
+    /// \returns True if the model was successfully added to the local cache,
+    /// and the model contains a model.config file.
+    public: virtual bool SaveModel(
+        const ModelIdentifier &_id,
+        const std::string &_data,
+        const bool _overwrite);
 
-      /// \brief Add a model from packed data to the local cache
-      /// \param[in] _id A completely populated ID
-      /// \param[in] _data Compressed content of the model
-      /// \param[in] _overwrite Overwrite model if already exists.
-      /// \returns True if the model was successfully added to the local cache,
-      /// and the model contains a model.config file.
-      public: virtual bool SaveModel(
-          const ModelIdentifier &_id,
-          const std::string &_data,
-          const bool _overwrite);
+    /// \brief Add a world from packed data to the local cache
+    /// \param[out] _id A completely populated ID
+    /// \param[in] _data Compressed content of the world
+    /// \param[in] _overwrite Overwrite world if already exists.
+    /// \returns True if the world was successfully added to the local cache
+    public: virtual bool SaveWorld(
+        WorldIdentifier &_id,
+        const std::string &_data,
+        const bool _overwrite);
 
-      /// \brief Add a world from packed data to the local cache
-      /// \param[out] _id A completely populated ID
-      /// \param[in] _data Compressed content of the world
-      /// \param[in] _overwrite Overwrite world if already exists.
-      /// \returns True if the world was successfully added to the local cache
-      public: virtual bool SaveWorld(
-          WorldIdentifier &_id,
-          const std::string &_data,
-          const bool _overwrite);
-
-      /// \brief Internal data.
-      private: std::shared_ptr<LocalCachePrivate> dataPtr;
-    };
-  }
-}
+    /// \brief Internal data.
+    private: std::shared_ptr<LocalCachePrivate> dataPtr;
+  };
+}  // namespace gz::fuel_tools
 
 #ifdef _MSC_VER
 #pragma warning(pop)
 #endif
 
-#endif
+#endif  // GZ_FUEL_TOOLS_LOCALCACHE_HH_
