+++ src/gz.cc
@@ -34,11 +34,6 @@
 #include <gz/msgs/fuel_metadata.pb.h>
 #include <gz/msgs/Utility.hh>
 
-#ifdef _WIN32
-// DELETE is defined in winnt.h and causes a problem with REST::DELETE
-#undef DELETE
-#endif
-
 #include <algorithm>
 #include <chrono>
 #include <deque>
