+++ src/gz_src_TEST.cc
@@ -51,6 +51,8 @@ class CmdLine : public ::testing::Test
     // instead of on teardown leaves the folder intact for debugging if needed
     common::removeAll(testCachePath);
     ASSERT_TRUE(common::createDirectories(testCachePath));
+    ASSERT_TRUE(common::createDirectories(
+        common::joinPaths(testCachePath, "fuel.gazebosim.org")));
     setenv("GZ_FUEL_CACHE_PATH", this->testCachePath.c_str(), true);
   }
 
@@ -85,7 +87,7 @@ TEST_F(CmdLine, ModelListFail)
 
   EXPECT_NE(this->stdOutBuffer.str().find("Invalid URL"),
       std::string::npos) << this->stdOutBuffer.str();
-  EXPECT_TRUE(this->stdErrBuffer.str().empty());
+  EXPECT_TRUE(this->stdErrBuffer.str().empty()) << this->stdErrBuffer.str();
 }
 
 /////////////////////////////////////////////////
